pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]

[project]
authors = [
    {email = "anyaagarenko@gmail.com", name = "Anya Agarenko"},
]
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-filter (==25.1)",
    "django-split-settings (>=1.3.2,<2.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-spectacular (==0.28.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest-freezegun (==0.4.2)",
    "pytz (>=2025.2,<2026.0)",
    "setuptools (>=80.9.0,<81.0.0)",
]
description = ""
name = "django-iw"
readme = "README.md"
requires-python = ">=3.11,<4.0"
version = "0.1.0"

[tool.poetry.group.dev.dependencies]
mixer = "^7.2.2"
pytest-django = "^4.11.1"
ruff = "^0.11.12"
toml-sort = "^0.24.2"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
addopts = ["--no-migrations"]
filterwarnings = [
    "ignore:distutils Version classes are deprecated",
    "ignore:pkg_resources is deprecated as an API",
]
pythonpath = "src"

[tool.ruff]
line-length = 160
target-version = "py311"

[tool.ruff.lint]
fixable = ["ALL"]
ignore = [
    "ANN001",  # missing type annotation for function argument `{}`
    "ANN002",  # missing type annotation for `{}`
    "ANN003",  # missing type annotation for `{}`
    "ANN201",  # missing return type annotation for public function `{}`
    "ANN202",  # missing return type annotation for private function `{}`
    "ANN204",  # missing return type annotation for special method `{}`
    "ANN205",  # missing return type annotation for staticmethod `{}`
    "ANN206",  # missing return type annotation for class method `{}`
    "ANN401",  # dynamically typed expressions (typing.Any) are disallowed in `{}`
    "COM812",  # trailing comma missing
    "CPY001",  # missing copyright notice at top of the file
    "D100",  # missing docstring in public module
    "D101",  # missing docstring in public class
    "D102",  # missing docstring in public method
    "D103",  # missing docstring in public function
    "D104",  # missing docstring in public package
    "D105",  # missing docstring in magic method
    "D106",  # missing docstring in public nested class
    "D107",  # missing docstring in `__init__`
    "D203",  # 1 blank line required before class docstring
    "D205",  # 1 blank line required between summary line and description
    "D213",  # multi-line docstring summary should start at the second line
    "D415",  # first line should end with a period, question mark, or exclamation point
    "FBT002",  # boolean default positional argument in function definition
    "ISC001",  # implicitly concatenated string literals on one line
    "RUF001",  # string contains ambiguous letter
    "RUF012",  # mutable class attributes should be annotated with `typing.ClassVar`
]
preview = true
select = ["ALL"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.per-file-ignores]
"*/migrations/*" = [
    "RUF012",  # mutable class attributes should be annotated with `typing.ClassVar`
]
"tests/*" = [
    "S101",  # use of `assert` detected
]

[tool.tomlsort]
all = true
in_place = true
sort_first = ["tool.poetry"]
spaces_before_inline_comment = 2
spaces_indent_inline_array = 4
trailing_comma_inline_array = true
